Refactored uuid_util and gateway_util into a `utils` module (#3689)
* partial progress

* partial progress

* finished refactoring

* fixed clippy
