+++ clients/rust/src/lib.rs
@@ -17,12 +17,12 @@ pub use tensorzero_core::endpoints::optimization::LaunchOptimizationParams;
 pub use tensorzero_core::endpoints::optimization::LaunchOptimizationWorkflowParams;
 use tensorzero_core::endpoints::optimization::{launch_optimization, launch_optimization_workflow};
 use tensorzero_core::endpoints::stored_inference::render_samples;
-pub use tensorzero_core::gateway_util::setup_clickhouse_without_config;
-use tensorzero_core::gateway_util::setup_postgres;
 use tensorzero_core::http::TensorzeroHttpClient;
 use tensorzero_core::inference::types::stored_input::StoragePathResolver;
 pub use tensorzero_core::optimization::{OptimizationJobHandle, OptimizationJobInfo};
 use tensorzero_core::stored_inference::StoredSample;
+pub use tensorzero_core::utils::gateway::setup_clickhouse_without_config;
+use tensorzero_core::utils::gateway::setup_postgres;
 use tensorzero_core::{
     config::Config,
     endpoints::{
@@ -33,7 +33,7 @@ use tensorzero_core::{
         validate_tags,
     },
     error::{Error, ErrorDetails},
-    gateway_util::{setup_clickhouse, GatewayHandle},
+    utils::gateway::{setup_clickhouse, GatewayHandle},
 };
 use thiserror::Error;
 use tokio::{sync::Mutex, time::error::Elapsed};
@@ -1321,7 +1321,7 @@ impl Client {
     }
 
     #[cfg(any(feature = "e2e_tests", feature = "pyo3"))]
-    pub fn get_app_state_data(&self) -> Option<&tensorzero_core::gateway_util::AppStateData> {
+    pub fn get_app_state_data(&self) -> Option<&tensorzero_core::utils::gateway::AppStateData> {
         match &*self.mode {
             ClientMode::EmbeddedGateway { gateway, .. } => Some(&gateway.handle.app_state),
             ClientMode::HTTPGateway(_) => None,
