+++ tensorzero-core/src/endpoints/dynamic_evaluation_run.rs
@@ -13,8 +13,8 @@ use crate::{
     db::clickhouse::{escape_string_for_clickhouse_literal, ClickHouseConnectionInfo},
     endpoints::validate_tags,
     error::{Error, ErrorDetails},
-    gateway_util::{AppState, AppStateData, StructuredJson},
-    uuid_util::{
+    utils::gateway::{AppState, AppStateData, StructuredJson},
+    utils::uuid::{
         compare_timestamps, generate_dynamic_evaluation_run_episode_id, validate_tensorzero_uuid,
         DYNAMIC_EVALUATION_THRESHOLD,
     },
