+++ tensorzero-core/src/endpoints/status.rs
@@ -1,6 +1,6 @@
 use crate::{
     db::HealthCheckable,
-    gateway_util::{AppState, AppStateData},
+    utils::gateway::{AppState, AppStateData},
 };
 use axum::debug_handler;
 use axum::extract::State;
@@ -64,8 +64,8 @@ mod tests {
     use std::sync::Arc;
 
     use crate::config::Config;
-    use crate::gateway_util::GatewayHandleTestOptions;
     use crate::testing::get_unit_test_gateway_handle_with_options;
+    use crate::utils::gateway::GatewayHandleTestOptions;
 
     use super::*;
