+++ tensorzero-core/tests/e2e/howdy.rs
@@ -16,10 +16,10 @@ use tensorzero_core::db::clickhouse::migration_manager::RunMigrationManagerArgs;
 use tensorzero_core::db::clickhouse::test_helpers::get_clickhouse;
 use tensorzero_core::db::clickhouse::ClickHouseConnectionInfo;
 use tensorzero_core::db::postgres::PostgresConnectionInfo;
-use tensorzero_core::gateway_util::GatewayHandle;
 use tensorzero_core::howdy::{get_deployment_id, get_howdy_report};
 use tensorzero_core::http::TensorzeroHttpClient;
 use tensorzero_core::inference::types::TextKind;
+use tensorzero_core::utils::gateway::GatewayHandle;
 use tokio::time::Duration;
 
 #[tokio::test(flavor = "multi_thread")]
